+++ lib/fluent_result.dart
@@ -1,5 +1,6 @@
 library fluent_result;
 
-export './src/result.dart';
-export './src/result_error.dart';
-export './src/result_of.dart';
+export 'src/extensions.dart';
+export 'src/result.dart';
+export 'src/result_error.dart';
+export 'src/result_of.dart';