+++ test/convert_results_test.dart
@@ -1,14 +1,13 @@
 // ignore_for_file: avoid_redundant_argument_values
 
 import 'package:fluent_result/fluent_result.dart';
-import 'package:fluent_result/src/result_error.dart';
 import 'package:test/test.dart';
 
 import 'helpers.dart';
 
 void main() {
   group('Convert success Result', () {
-    test('to another succes Result', () {
+    test('to another success Result', () {
       final result = getRandomCustomer(id: 1);
 
       expect(result.isSuccess, true);
@@ -22,7 +21,7 @@ void main() {
 
   group('Convert fail Result', () {
     test('to another fail Result', () {
-      final result = ResultOf.fail<User>(const ResultError('user not found'));
+      final result = ResultOf.withError<User>(ResultError('user not found'));
 
       expect(result.isFail, true);