+++ test/custom_result_error_test.dart
@@ -9,7 +9,7 @@ import 'helpers.dart';
 void main() {
   group('Result Error', () {
     test('is fail with custom error', () {
-      final result = Result.fail(const CustomerNotFound(customerId: 33));
+      final result = Result.withError(CustomerNotFound(customerId: 33));
       expect(result.isFail, true);
       expect(result.isSuccess, false);
       expect(result.error is CustomerNotFound, true);
@@ -26,7 +26,7 @@ void main() {
   test('Many errors', () {
     final err1 = CustomerNotFound(customerId: 1);
     final err2 = InvalidPasswordError('The password 123456 is invalid');
-    final res = Result.fails([err1, err2]);
+    final res = Result.withErrors([err1, err2]);
 
     res.isFail.should.beTrue();
 
@@ -37,9 +37,9 @@ void main() {
     res.contains<InvalidPasswordError>().should.beTrue();
   });
 
-  test('Caontains error', () {
+  test('Contains error', () {
     final err1 = CustomerNotFound(customerId: 1);
-    final res = Result.fail(err1);
+    final res = Result.withError(err1);
 
     res.contains<CustomerNotFound>().should.beTrue();
     res.get<CustomerNotFound>().should.not.beNull();
@@ -55,7 +55,7 @@ void main() {
   test('Add error', () {
     final err1 = CustomerNotFound(customerId: 1);
     final err2 = InvalidPasswordError('The password 123456 is invalid');
-    final res = Result.fail(err1);
+    final res = Result.withError(err1);
 
     res.add(err2);
   });
@@ -63,11 +63,11 @@ void main() {
   test('Equal results', () {
     final err1 = CustomerNotFound(customerId: 1);
     final err2 = InvalidPasswordError('The password 123456 is invalid');
-    final res1 = Result.fails([err1, err2]);
+    final res1 = Result.withErrors([err1, err2]);
 
     final err11 = CustomerNotFound(customerId: 1);
     final err22 = InvalidPasswordError('The password 123456 is invalid');
-    final res2 = Result.fails([err11, err22]);
+    final res2 = Result.withErrors([err11, err22]);
 
     res2.should.be(res1);
   });